Move teehistorian UUIDs into engine
