+++ src/engine/shared/global_uuid_manager.cpp
@@ -1,4 +1,5 @@
 #include "protocol_ex.h"
+#include "teehistorian_ex.h"
 #include "uuid_manager.h"
 
 #include <engine/uuid.h>
@@ -7,6 +8,7 @@ static CUuidManager CreateGlobalUuidManager()
 {
 	CUuidManager Manager;
 	RegisterUuids(&Manager);
+	RegisterTeehistorianUuids(&Manager);
 	RegisterGameUuids(&Manager);
 	return Manager;
 }
