+++ src/game/server/teehistorian.cpp
@@ -9,7 +9,7 @@ static const CUuid TEEHISTORIAN_UUID = CalculateUuid(TEEHISTORIAN_NAME);
 static const char TEEHISTORIAN_VERSION[] = "2";
 
 #define UUID(id, name) static const CUuid UUID_ ## id = CalculateUuid(name);
-#include <game/teehistorian_ex_chunks.h>
+#include <engine/shared/teehistorian_ex_chunks.h>
 #undef UUID
 
 enum
